refactor(print-format-builder): extract useLayoutHistory composable
